pyproject.toml
[project]
name = "viagoscrap"
version = "0.1.0"
description = "Scraper Viagogo en Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "playwright>=1.52.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "apscheduler>=3.10.4",
  "httpx>=0.27.0",
]

[project.scripts]
viagoscrap = "viagoscrap.cli:main"
viagoscrap-web = "viagoscrap.webapp:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
